requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0  # PyTorch Image Models

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0  # Advanced augmentations

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress & Logging
tqdm>=4.65.0
tensorboard>=2.14.0

# API Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ONNX Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Grad-CAM
pytorch-grad-cam>=1.4.0
